comment added; usage added
