--- BUILD
@@ -1,7 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_binary(
-  name = "find_expected_governor",
-  srcs = ["find_expected_governor.cc", "parse_forest.pb.h", "expected_governor.h"],
-  deps = ["@protobuf//:main"],
-)
