+++ README.md
@@ -1,2 +1,15 @@
 # expected_governor
-find expected governor given parse forest
+@author ytdu (duyantao2009@gmail.com)
+
+Find the expected governors given a parse forest.
+
+Usage:[
+-input <input path (Default=test.predict)>
+ -output <output path (Default=stdout)>
+ -rule_path <binary rule path (Default=data/tcrf_rule)>
+ -head_rule_path<binary headrule path (Default=data/binary_headrules)>
+ -debug
+]
+
+Reference:
+Helmut Schmid and Mats Rooth: Parse forest Computation of Expected Governors. ACL'2001
