--- WORKSPACE
@@ -1,5 +1,0 @@
-new_local_repository(
-    name = "protobuf",
-    path = "/usr/local",
-    build_file = "protobuf.BUILD",
-)
